% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airports-package.R
\docType{package}
\name{airports-package}
\alias{airports}
\alias{airports-package}
\title{airports: Data on Airports}
\description{
Geographic, use, and property related data on airports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OpenIntroStat/airports}
  \item Report bugs at \url{https://github.com/OpenIntroStat/airports/issues}
}

}
\author{
\strong{Maintainer}: Mine Çetinkaya-Rundel \email{cetinkaya.mine@gmail.com} (\href{https://orcid.org/0000-0001-6452-2420}{ORCID})

}
\keyword{internal}
